feat: multiple user agent support
This commit will use curl useragent for facebook and Instagram but will use random useragebts for other sites

The aim is to reduce the preview endpoint from being marked as spam. Cloudflare for instance will not play nicely with curl UA.
Facebook/IG on the other hand will not play nicely with non-curl. Or maybe I should test this for the preview endpoint