ft: add CORS
Add CORs to fix the current issue with previewing urls. This however is
not good enough because in is using a wilcard. In future, this should
not be global. Auth should only be available in the internal routes like
/api.

It might also be a good idea to write the CORs logic manually.
For now this is good enough.